chatroom project in Django and Rest API,
chat room application : features:
a home page with all topics. create room and in the room logged in  member can chat , also we have participants sections which show who all are participating in the chat room.
and we have activity area in the home page were user can see all the members and their rooms and their activities..